BUG #13466. Adding support for UserTranform on representations.
Added a mechanism to specify a 4x4 matrix as the UserTranform for a geometry
representation. Python scripts can now set the UserTranform (results in calling
vtkActor::SetUserTransform()), using the UserTranform property which is expected
to be a 16 element double corresponding to the 4x4 matrix.

Along with vtkGeometryRepresentation, updating vtkSelectionRepresentation and
vtkDataLabelRepresentation to accept UserTransform so that selections and labels
on selections show up correctly when a UserTransform set on the
vtkGeometryRepresentation.