SciberQuestToolKit -- intel includes
gcc intinsics are erroneously found by the intel compiler, try to
prevent this by finding and including intel's directory explicitly.
if we can't locate the header then a warning is issued, the user
may have to set the path if there are problems. There may not be
problems if the intel compiler searches its directories bofore
gcc's or gcc's copy of the intriniscs don't conflict with intels.