Fixed client-server AMR volume rendering issue.
When locally rendering, the resampled volume was getting reset when a streamed
piece was updated. This was because the
vtkResampledAMRImageSource::UpdateResampledVolume() was checking the AMR data's
mtime to determine if we needed to reinitialize the datastructure. That's
incorrect. Since in client-server mode the client is delivered the AMR for every
update, we end up with a new mtime and hence we were always reinitializing the
datastructure. Fixed that. vtkAMRStreamingVolumeRepresentation::RequestData()
takes care of initializing the vtkResampledAMRImageSource when the data changes
anyways.