Fixing tests for static builds of ParaView.
The GUI automatically loads all plugins for static paraview builds
so we I created a separate XML GUI test script that doesn't try
to load the coprocessing script generator plugin. Another problem
though is that with all of the other plugins automatically loaded
the outputted script will have extra properties set that aren't
available when the "simulation" is run which causes the full workflow
test to fail so that isn't run with static builds, for now at least.

Change-Id: I25455e7ba366bf4546beafac45393a12a55c8573