BUG #13482. Fix outlines for reclinear and structured grids.
Outline for vtkStructuredGrid and vtkRectilinearGrid was not working correctly
in parallel. We would see internal outlines. The issue was introduce by VTK 6.0
API changes. Since the data object no longer provides pipeline information and
hence no WHOLE_EXTENT information is available, we have to explicitly create
vtkPVTrivialProducer to pass the whole_extent correctly to the internal filters.

Adding a new test (DataOutlines) to test this functionality.