Fix ResetDisplay for charts.
Sometimes ResetDisplay would have no effect on charts. As soon as one interacted
with the chart and called ResetDisplay again, it would work. The problem was
that if the *AxisRange properties were already unset, then they would not be
pushed and hence the view would not reset the axis. This fixes that issue by
ensuring that the *AxisRange properties are pushed.