Added mode to resample volume based on view frustum for AMR.
Added mode to determine the placement of the uniform grid based on the AMR
bounds and the view frustum. This borrows from the code in vtkAMRVolumeMapper.
When "Resampling Mode" is set to "Using View Frustum" for AMR volume rendering,
this new mode is used. Currently, it will not work in client-server mode with
local rendering since the camera is not valid on the server-side in that case.

Change-Id: I84f4577fa567a428b0be235a86f188a906edf6a4