Working on cleanup connection on Catalyst disconnect.
When Catalyst disconnects from ParaView, we need to clean up the Catalyst
session show to the user. The user is prompted to notify that Catalyst
disconnected (in future we may add support to save state) and then the Catalyst
session is cleaned up.

We don't remove any extracts that the user was visualizing. Those can be
manually deleted if no longer needed.

Change-Id: Ie018e763057a82262e80e806f740006d11570c45