Fixed BUG #13525. Avoid rendering pipeline resets when playing animation.
When playing animation, display lists are always discarded even if pipeline data
didn't change with time. Although the geometry filter and the vtkPVCacheKeeper
in vtkGeometryRepresentation didn't re-execute, the data was still treated as
modified. This was due to the logic in vtkPVDataDeliveryManager::SetPiece(). The
code was implemented to look at the DataTime provided by the pipeline. That was
done so that we could detect the change in data object when representation type
changes. However, the same can be done simply by comparing the data-object
pointer. Doing that addresses this bug as well.

Note however, in client-server mode, we still would force delivery for all
representations when time changes irrespective of whether the data really
changed.

Change-Id: I1b41778a5e3fcfa041ff48860c5311245a5182a7