Cleaned up API to control ordered compositing process order.
When ordered compositing is enabled, ParaView determines the process order by
partitioning the data using PKdTree. However, in some cases, like when volume
rendering image-data, we need to override the partitioning using the partitions
from the image data itself. This commit cleans up the API for proving such
information to the view from a representation. In past it was tightly coupled
with vtkImageData producer. This change makes it possible to just pass the
meta-data about extent-translator to use, origin, spacing, extents etc. to
generate the partitioning. This will make it possible for non-image data
representations to provide this meta-data without having to create a
vtkImageData.

This removes obsolete vtkBSPCutsGenerator. Also vtkKdTreeManager now always
regenerates the KdTree. The checks to avoid regenerating unless the input
changed have been removed.

Change-Id: I469f54ba931fd64f66097d0ea579e9d8652d9bfc