Add ability to support volume render vtkStructuredGrid.
vtkStructuredGridVolumeRepresentation extends
vtkUnstructuredGridVolumeRepresentation to add ability to volume render
vtkStructuredGrid by converting it to vtkUnstructuredGrid internally.
Additionally, for  non-curved grids i.e. in cases when bounding boxes of
non-intersecting extents don't intersect (or intersect negligibly), we can use
the partitioning provided by the producer (using vtkExtentTranslator) to
determine the KdTree for ordered compositing in parallel (similar to what we do
for vtkImageData).

Change-Id: I55cf1817045ebadc53b7e7503a07589377c397e0