Fixed BUG #13469. Fix clipping of cube-axes when parallel rendering.
Cube-axes labels were getting clipped with parallel rendering was used. This was
because the CubeAxes would only report its data bounds to the renderer and not
the actually rendered bounds. vtkIceTCompositePass uses those bounds to decide
what pixels need to be composited and that was resulting in the labels being
ditched. Fixed that by using a heuristic to scale the bounds reported by
CubeAxes. Since it has no rendering side-effects besides a few extra pixels
being composited, we opt for this heuristic based approach rather than anything
more concrete.

Change-Id: I41ee70919349a15f08ef14a40d6073bc9f64168c