BUG #13326. Fixed segfault in vtkPExtractHistogram in parallel.
Cleaned up the logic to determine the range for the bins. The logic was
confusing and hence must have missed some corner case which was causing
segfaults. This undoes an earlier commit to communicate the name of for the
"bin_extents" array since that array name is no longer picked based on the array
that was chosen, it doesn't made much sense to communicate that information
anyways.

Change-Id: Ifcf74b8e541dfebb06b4d204c0edff70dd8372f2