BUG #12290. Fix loading of state file with histograms.
Old state files for charts often saved the CompositeDataSetIndex as 0. In that
case we need to handle that by passing the entire data set for plotting. This
fix addresses that use-case. This fixes the state loading issue reported in the
BUG (although the bug was probably referring to an old issue that's already been
addressed, since the plot renders correctly after this fix).

Change-Id: Id4f21319e3c89c3d1fa94f8eb82334ace83e90d1