Make sure Alt+A is the ONLY Apply button shortcut across all platforms
On Mac, naming the button &Apply was not adding the Alt+A shortcut where as other platform will do it automatically.
And binding it to Ctrl+A or Commad+A was also bad specially on TextField as Ctrl+A means "Select all" which was preventing the apply to happen at all.

So fixed that by adding the Alt+A shortcut to the button as well as keeping the &Apply text label.

Change-Id: I72f6a409d0931921e36dd48c14eb042edea182db