Fix SciberQuest test with MPI in builtin mode
This fixes an issue in which, when the SciberQuest plugin is
loaded and ParaView is compiled with MPI enabled but running in
builtin mode, the vtkSQBOVReaderBase class will emit an error
message. Now we check for MPI initialization in CanReadFile() so
that SetFileName() will not be called when MPI is not initialized.

Change-Id: I5c102d8e0555984530ca5006ff59233cd5875f31