Add the MobileRemoteControl plugin
The mobile remote control plugin allows a mobile device to view
the ParaView scene and control the camera.  The plugin uses a
threaded server so that communication with the mobile device does
not block the ParaView GUI.

The mobile device client is implemented in cross platform C++
and will be available for iOS and Android.

Change-Id: Ie914bcf19d85e580bf06e568a9c9363cd6f2c19a