Disable MobileRemoteControl plugin by default
This changes the MobileRemoteControl plugin to be disabled
by default due to it causing dahsboard failures on static
builds.

If/when the issues are resolved we can set it back to enabled
by default. In the mean time users can enabled it via cmake by
checking the PARAVIEW_BUILD_PLUGIN_MobileRemoteControl option.

Change-Id: I9973ccce68488bc2060817ea45b50e0bbb5cd139