Rewrite the pqVRPNEventListener to avoid mutexes entirely.
This also fixes the usage of QThread in the VRPN listening
code (subclassing QThread is not the preferred way to use it:
http://blog.qt.digia.com/blog/2010/06/17/youre-doing-it-wrong/ )

Instead of locking during listen() to protect access to the
connection container, this patch uses the thread's event loop to
ensure that the container can only be access from a single thread
at a time. This is accomplished by adding/removing connections in
slots that are handled in the same thread that is continuously
listening on the sockets. The incoming request to add/remove a
connection interrupts the listening loop and listening resumes
once the request is handled.

Change-Id: I74c1e4811f4c39d55d3a86e30cd0bb109c0b9f69