Add a "use continuous range" option to threshold filter.
If off, the vertex scalars are treated as a discrete set. If on, they
are treated as a continuous interval over the minimum and maximum. One
important "on" use case: When setting lower and upper threshold equal
to some value and turning AllScalars off, the results are cells
containing the iso-surface for that value.

Change-Id: I975b36e90f0697ddc2a9b20e459d41347a21746e