BUG #13231. Fixed spreadsheet view when data was deleted.
Spreadsheet view would not update to show 0 rows/columns when the data source
being shown in the view was deleted. This caused misleading text display. This
fixes that issue. vtkSpreadSheetView needed to report that the data had changed
otherwise the UI would not reset it self resulting in the bug reported.

Change-Id: I46f81c050b6445bc4c577dc0280cb9169a45b661