Add toolbar for subset of Data Analysis filters.
This extends ParaView's XML menu item syntax to include a new
attribute for proxies: if the `omit_from_toolbar` attribute is
present in a `Proxy` tag and the proxy's containing `Category`
has `show_in_toolbar`, then the proxy will appear in submenu
for the category but not in the toolbar.

This allows us to provide a Data Analysis toolbar that does not
repeat the Calculator filter, which also appears in the Common
category. Furthermore, we do not need to provide icons for
less-frequently-used menu items such as Plot On Sorted Lines.

Change-Id: I9dc11e299b9329f51283a829d4c08da6ac89bc11