Fix bug 10968 ("freeze" query selections).
This changeset, along with a matching change to VTK, provides
users with a way to convert query selections into index-based
selections -- thus "freezing" the selected cells to one particular
timestep's attribute values.

In order to accomplish this, the `vtkPConvertSelection` class
passes an instance of `vtkPVExtractSelection` to its superclass for
use when converting geometric and attribute-based selections into
indexed-based selections.

The `vtkPConvertSelection` class is changed to partially support
the `PreserveTopology` flag employed by `vtkConvertSelection`; it
generates the "vtkInsidedness" array on its output. This in turn
requires changes to `vtkPythonExtractSelection` to pass its mask
array instead of subsetting the input dataset.

Finally, note that this change moves `vtkPConvertSelection` to the
`vtkPVClientServerCoreCore` library to avoid complicating dependencies.

Change-Id: Id9f1ebd1ea3afe9b595c58d54a49bfbb9abbdedf