BUG #13739. Fixes issues when user clicked on the check-box itself.
Toggling the check-state of multiple items wasn't working correctly when the
user clicked on the "checkbox" itself, rather than the name of the item. This
fix address that issue. We make each of the items non-checkable. Thus Qt no
longer deals with the check-box toggling. This results in the clicks on the
"check-box" to be treated similar to clicks on the item's name. We have code
that handles that already and hence everything works as expected. Now the
perceived behaviour is identical whether the user click on the checkbox or its
name in the variable selection widget.

Change-Id: I30fb162788a8cf47214f72b92a91cbb1db6c9ff8