Fix bug 10958 (plot query-selections over time).
Query selections were not supported by the "Plot Selection Over Time"
filter because vtkPExtractArraysOverTime and vtkExtractArraysOverTime
created an instance of vtkExtractSelection (which does not currently
handle query selections) to obtain the selected points/cells at each
time step.  This patch provides an instance of vtkPVExtractSelection,
which uses ParaView's Python interpreter to process the query selection.

The underlying VTK classes have also been modified to compute
statistics on the selection and report those instead of creating
a vtkTable time-history for each point or cell in the selection.
Statistics are always reported for query selections and an option
is provided in the ParaView UI to obtain statistics for other
selections.

Change-Id: Iab7f9bceacb9a5e07fc6b5179b862c6981e5e392