Enable the vtkWindBladeReader for serial server runs.
VTK was modified to have a serial and a parallel version where the
parallel version relies on MPI IO. We use the factory to create
the proper one whether we're running in serial or paralle.
Before there was just a parallel version and it was only enabled
in the ParaView GUI when pvserver was run with multiple processes.