BUG #13799. Variables not showing curretly for Comparative View.
The problem was that the representations in each view were being updated using
repr->Update() instead of simply letting the view update the representations as
needed. Since vtkCompositeRepresentation::Update() no longer forwards the update
request to internal representations, it meant that the UI ended up updating
using old state of the representations -- hence the missing variables. This
fixes that issue.

Change-Id: Ia25e7dd8f4d5c0c368692894003567eb8bf79043