Fixed Overlay mode in Comparative View.
When in Overlay mode, all representations were simply showing the first
timestep. This is because when the view was updated for (0, 0), it would update
all representations in it, including those for other locations in the
comparative grid. The caches for all those representations would be filled and
subsequence updates would  have no effect. This addresses that issue by clearing
caches for representations at (i, j) location just before updating the
representations at (i, j) location.

Added a test for the same (ComparativeOverlay).

Change-Id: I05493e4258e54a7dc9b4554c56f40f8271e22f21