Fixing UndoRedo3 tests.
A 'fix' committed in 81fe31a6 was resulting in the 3DWidget show being shown
when the creation of a clip filter, for example, was undone and redone. This
patch addresses that issue. I believe the widgets for a proxy were destroyed and
recreated in the past which is no longer the case and hence the 'fix' is not
needed.

Change-Id: Icedeae3af16e2e0b1bb5c578a1558343ccd08837