Add ability to export/import VR plugin configuration from file.
This allows a minimial pvsm-style file to be written that contains
the VR plugin state.

The "proxyName" attribute has been added to the <Style ...> tags
that can be used to lookup a proxy by name, since the pvvr file
won't contain any information to map the proxy id into a usable
object.

Change-Id: I08891cd9664f70bfad5b32dd1b00fa451158c837