Fix viewport size.
When placing a viewport within a fullscreen div, the viewport ended up being
resized periodically to an incorrect size. This fixes that issue. The problem
seemed to be due the render request happening at a mildy stale point in time
thus the image receieved from the server could be slightly different than the
view. Since the canvas ended up being resized to the image size, we ended up
distorting the viewport size.

Change-Id: I3ddda0071355f7c819fa67341e5f6eedd545e1f8