BUG #13855. Fixed segfaults when loading state.
Fixed segfaults when loading state with charts and current time > 0. The problem
was that the chart would render (due to change in time) before the QVTKWidget
was created. That resulted in the window popping up and then crashin when the
QVTKWidget was getting created at a later point.

Fixed by ensuring that the QVTKWidget is setup soon after the proxy is
registered and detected by the GUI.

Change-Id: I2124d6c5243b4a20131b53cd708c20724dff8cf7