Distinguish Plugin Readers.
VisItBridge Analyze reader and AnalyzeNIfTIReaderWriter plugin reader
had the same file_descriptions. Either a '(Visit)' could be appended to
the Visit reader of a '(Plugin)' could be appended to the plugin
reader. We picked the later.

This commit goes a step further and appends all the Plugin readers with
'(Plugin)'.

Change-Id: If0c48cb18b135d622402432987f422cd09acc6d7