Fix AUTOINIT issues when building external plugins.
VTK_USE_FILE adds definitions for ${module}_AUTOINIT for all enabled modules.
This is okay for VTK, with ParaView, AUTOINIT is not useful since one needs
to init the CS streams separately. Also use AUTOINIT is defined, any application
needs to link against all enabled VTK modules which ends up being a very long
list and hard to keep up-to-date. We over come this issue by simply not setting
the ${module}_AUTOINIT definies.  So we get the current COMPILE_DEFINITIONS on
the directory and remove references to AUTOINIT.

Change-Id: I2345223af5588e0b9979f89dc7fa628cf6782278