BUG #13927. Fixed borders on text widgets.
Since the shared-render window had so size set initially, the
vtkTextRepresentation ended up incorrectly computing the placement for the text
and hence the state for the border visibility in batch mode. This fixes that
issue by ensuring that we setup a default size for the shared window as well.
The size gets updated when rendering happens to the actual size, so we don't
need to be extra smart about the deafult window size.

Change-Id: I1d9eb5dbc689aec838539fe9b0fa152db6b31801