We now don't sort arrays when using vtkCompleteArrays.
When parallel writers are used, process 0 writes out the
meta-data file but it can't do it properly if it doesn't have any
points or cells so vtkCompleteArrays shares this information.
The other modified classes automatically sort the array information
but we don't want that in this case so I set it to turn sorting off.

Change-Id: I1e3c0affaabfff771aebd7828f69cd1e86db5065