BUG #13933. Allow non-present arrays in Python scripts.
"ColorArrayName" property no longer checks the validity of the array specified
so long as the array is specified as  a tuple (<attribute>, <name>). It will
check (and fail if array not found) when the user simply specifies the array
name as a string.

Updated Python trace to ensure that for "ColorArrayName" we indeed trace out a
tuple rather than just the arrayname.

Also cleaned up vtkPVOptions to expose publicly accessed enum. Updated
servermanager to use the enum rather than the hex values.

Change-Id: I219e83ba58b825e40bb4c5a32b69207ff57d9d3c