BUG #13956. Fix duplicate execution of filters.
The problem is that the SelectionRepresentation doesn't ask for ghost cells
while the GeometryRepresentation does ask for ghost cells (BUG 0011811). Since
SelectionRepresentation is added before the GeometryRepresentation, it updates
the pipeline gets updated with 0 and then 1 ghost-cell request, hence the double
execution.

The GeometryRepresentation, which is the primary data-representation should
always be updated before the Selection Representation. So to fix it, we changed
the order in which the representations get added.

Change-Id: Iaadf8e952a99fddc105d6e41d09f71e4a2240d7b