Refactoring Properties Panel (pqPropertiesPanel).
This commit refactors pqPropertiesPanel for better usability and visual
appearance. The changes can be summarized as follows:

- Added a new class, pqProxyWidget, which creates the 'auto-generated' panel for
any proxy. pqPropertiesPanel now uses this class to create the panels for
source/display.

- Property group were extended to add ability to group properties on non-exposed
properties (as demonstrated by Wavelet source). The visual appearance of
property groups was also changed. Also fixed leaks in ServerManager relating to
groups and new property attributes that were added in past for the panels.

- pqPropertyWidgetDecorator was added to provide a mechanism to add custom logic
to widgets which having to create new widget types. pqPropertyWidgetDecorator
can be used to control widget visibility, for example.

- Converted custom panels for for several filters to be non-custom i.e. simply
be auto-generated.

- Added mechanism for plugins to provide new pqPropertyWidget types for
properties, property groups or new pqPropertyWidgetDecorator types. Added an
example under Examples that demonstrates the use of the same.

- Added infrastructure to print debug messages when certain environment
variables are set (pqDebug).

- Removing obsolete SummaryPanel elements.

- Updated tests to work with these changes.

Change-Id: I49e82525ba666380b265a7f2075207dc2cd6f452