Adding a warning for a change that isn't backwards compatible.
We split apart vtkCPProcessor into vtkCPPythonProcessor and vtkCPProcessor.
vtkCPPythonProcessor does initialization for ParaView and Python for
Catalyst but vtkCPProcessor only does initialization for ParaView.
vtkCPProcessor did used to also do initialization for Python though
so we need to add in a warning/error that the code has changed if
they try to add in a vtkCPPythonScriptPipeline if only vtkCPProcessor
is being used, otherwise bad things will likely happen.

Change-Id: Ib09dd560c913b4281c789f4a2931663c7d083931