Move DataAnalysis toolbar.
For every category of toolbar that is added, ToolbarBreak is
applied. This causes a new toolbar to show up below the previous and
ends up leaving blank unused space in the toolbar region.

We want all toolbars loaded after the Common toolbar (including
DataAnalysis) to come up next to it and not below it. This means that we
apply a ToolbarBreak only before the Common toolbar and not afterwards.

Change-Id: I26c6e2b5faae9aeb6627610f71cb0ab0cd88f097