BUG #14046: Fixing issues with interaction modes and multiple views.
The use pqRubberBandHelper between multiple views was causing weird interaction
issues with selection modes, picking and such. Refactored the code to deprecate
pqRubberBandHelper. Instead, we now use pqRenderViewSelectionReaction to assist
with selection modes for the render views. Cleaned up usage of
pqRubberBandHelper to use the new refactored code.

Leaving pqRubberBandHelper for backwards compatibility. It should be removed in
future versions.