Tweak scalar bar actor.
This changes the default position of the annotations and
tick marks so that the minumum and maximum values appear
when the actor is in its default screen location.

This also carefully formats the min and max labels so that
they have a reasonable number of digits (3 or the number
required to distinguish the min and max, whichever is larger).

Change-Id: Ifc2605758b105a8aa7ddcbf9960c288a19da4e83