Overcome issues with trace.
When saving Python state, as we iterated over existing  proxies, ProxyProperty
(in servermanager.py) would register the proxies in this domain (since the GUI
didn't) and would cause the state to fail. Since this domain is not used by
ParaView, I am removing it for now. We need to cleanup how ProxyListDomains are
dealt with in ParaView in general.

Change-Id: I83cf213fac36f25b1ce01911c23b8d3c48e1e269