Cleanup text displays from Python.
Unless the Python shell is executing code, all Python outputs are now simply
shown in the OutputWindow (and terminal). Since Python adds "\n" appropriately
to the messages, extended pqOutputWindowAdapter/pqOutputWindow to add API to not
add extra "\n" as well as skip the echo to terminal (since vtkPythonInterpreter
does that already).

Change-Id: I869c025532faab4dfd5bf1011269ef1ea547cdea