Allowing co-processing scripts to be generated when nothing is outputted.
Before we checked if a writer proxy or view output was specified
and if neither were specified then we would prevent the user
from creating a script. Now we changed it so that there is only
a warning if no output is specified. Also, the we count that
if the user checked Live Visualization that something is being
outputted and we don't give the warning in this case either.
I also got rid of some trailing white space since that annoys me.

Change-Id: I48b507c93c984e45ac6652743ea25a52f0e3d18b