Fixes for modeless Find Data.
Rewrite `pqDataQueryReaction::showQueryDialog()` so that only a
single Find Data dialog gets instantiated, and do not call
`QEventLoop::exec()`.

Also, the dialog misbehaves when no source is selected in the
pipeline inspector: the list of field associations is not
correct and the arrays in the default query clause are always
point data (even though Cell is usually the default association).
This fixes the `pqQueryDialog` constructor by choosing the default
filter as the producer when NULL is passed.

Change-Id: I1566b58a568304077aaf2503861dd842414b3fa2