Change the behavior of pqPythonShell::pushScript() to split over newlines
Previously, you could not paste multiline statements into the ParaView
python console.  This commit changes the behavior of the pushScript()
function so that it splits the script code over the newline character
and pushes each line separately.

This does change the behavior of entering multiline statements: previously
pushScript() would attempt to detect the indent and start the next line
using the indent of the previous line.  This behavior has been removed.

The new behavior closely matches the behavior of the standard commandline
python interpreter.

Change-Id: I899ec69194a1cce3d82cefc2f13048aad8f81fec