Fix: in the programmable filter, delete the function before redefining it
This was a bug in the programmable filter.  If there was a syntax error
in the python script, then the python code that defined the RequestData
function (or other functions) would fail, and then the filter would
execute the previously defined version of the function.  The previously
defined version could have even been defined by a different programmable
filter.

In the future, the programmable filter could be refactored so that it
executes all code in a namespace (it could use the
vtkInteractivePythonConsole) which is cleaned up each time.  Currently,
it stores variables directly in the __main__ context.

Change-Id: Ieba0ca70396aa08fb7f8a8cc93aa4e2edc3f71c9