Use the python main context instead of the interactive console for trace.
The record trace and trace state features can execute in the __main__
python context using vtkPythonInterpreter::RunSimpleString() instead
of executing in the context of the interactive console owned by the
pqPythonShell.

I can't think of a reason to prefer one context instead of the other, but
using the __main__ context simplifies the code in pqPythonManager.
In either case, the only variables written to the context is the imported
smtrace module name and the __smtraceString str object.

Change-Id: I5239067c9d95a27f9a5867fd075b4bc146c4da76