Adding option to load plugin at command line.
Two options are added to the command line namely --test-plugin and
--test-plugin-path. These will be used exclusively for loading and
testing individual plugins in paraview.

--test-plugin option takes the name of the plugin and loads it. The
  default plugin load paths are used with this option.

--test-plugin-path append the provided path to the actual search path.

Also adding two new cmake functions i.e add_client_plugin_tests and
add_client_server_plugin_tests which takes in Plugin name as an
additional option and passes it along with the --test-plugin command
line.

We are changing multiple plugin tests to use this options instead of
loading it in the test script. We also get rid of all xml code to load
the plugins from the test scripts.

Plugins have to be loaded on both client and server. The server-manager
on the client side typically helps load the plugin from default
location. However if the plugins were loaded from a different location
then the plugin-path needs to be specified not only at the client side
but also on the server side. This commit also achieves this.

The add_client[_server]_plugin test is extended to allow
plugin_path. The test driver is changed slightly to parse more options
sent to the server.

Change-Id: I15c26748979b280edae817cd34ea198cfe4a7bfe