Fix logic to connect to renderserver nodes.
When running in data-server/render-server mode, the data-server wasn't getting
the information about the hostnames for the render-server nodes correctly and
hence always ended up connecting to "localhost". This fixes that. Now, the
information about the render-server nodes' hostnames is relayed to the
data-server correctly. The render-server nodes' hostnames are determine using
plaform specific API. At the same time, argument --hostname is provided to
override the platform specific hostname for the process (useful for debugging, as
well).

Change-Id: Idf7929b5a7cc5c800579e353418e0a9066f79f5e