BUG #14112: Fixed parallel surface selection with large number of elements.
Fixed BUG #1411. Surface selection was failing in parallel since there was
mismatch between the client and server processes when deciding whether they
needed to render additional passes for the ID. The additional passes are
required when the max-id of the elements per node are greater than 0xffffff.
Added logic to synchronize the element count among all processes right after the
first selection-pass so that the subsequent passes can make the correct decision
on all processes invovled.

Extended vtkPVSynchronizedRenderWindows to allow reduction operations with
MIN, MAX and SUM (previously, we only supported sum).

Change-Id: I9ddd43a80acdec7d1dd1d9cd98b3d138f1577f20