Fix for the client to wait before updating the pipeline.
For ParaView trying to connect to a Catalyst linked simulation,
it was possible that the client would try to update the pipeline
before the server has received the data extract from Catalyst.
This fixes that issue by having the server wait until it has
received all of the data objects and inserted them into
trivial producers before notifying the client that there is
another time step available.

Change-Id: I5f9d550d5f0e7a41a72ba607b97cdaa42dfd13f7