Fix bug when starting animation from middle
This fixes a bug that occurs when start an animation from the middle
of the time range when in loop mode. Previously the animation would
only loop from the middle to the end. Now it will properly loop from
the middle to the end and then restart from the beginning.

Change-Id: I8a11dc849a838f18dfc26b8beacc828fa770a1d1